+++ src/main/java/com/buuz135/industrial/proxy/CommonProxy.java
@@ -23,6 +23,7 @@
 package com.buuz135.industrial.proxy;
 
 import com.buuz135.industrial.proxy.event.FakePlayerRideEntityHandler;
+import com.buuz135.industrial.proxy.event.MobImprisonmentTamableEntitiesHandler;
 import com.buuz135.industrial.utils.explosion.ExplosionTickHandler;
 import com.hrznstudio.titanium.event.handler.EventManager;
 import net.neoforged.neoforge.common.NeoForge;
@@ -32,7 +33,7 @@ public class CommonProxy {
 
     public void run() {
         NeoForge.EVENT_BUS.register(new FakePlayerRideEntityHandler());
-
+        NeoForge.EVENT_BUS.register(new MobImprisonmentTamableEntitiesHandler());
         EventManager.forge(ServerTickEvent.Pre.class).process(ExplosionTickHandler::serverTick).subscribe();
     }